REF: misplaced sort_index tests (#33774)
